refactor(misc): make caput.misc a subpackage
