+++ caput/misc/__init__.py
@@ -1,15 +1,9 @@
 """A set of miscellaneous routines that don't really fit anywhere more specific."""
 
-import collections
 import importlib
-import logging
 import os
-import time
 
 import numpy as np
-import psutil
-
-logger = logging.getLogger(__name__)
 
 
 def vectorize(**base_kwargs):
@@ -265,7 +259,7 @@ class lock_file:
     Examples
     --------
 
-    >>> from . import memh5
+    >>> from .. import memh5
     >>> container = memh5.BasicCont()
     >>> with lock_file('file_to_create.h5') as fname:
     ...     container.save(fname)
@@ -332,7 +326,7 @@ def getfullargspec(*args, **kwargs):
     import warnings
 
     warnings.warn(
-        "This patch to support Python 2 is no longer needed and will be removed.",
+        "This patch to support Python 2 is no longer needed and will be removed after December 2021.",
         DeprecationWarning,
     )
 
@@ -362,109 +356,3 @@ def import_class(class_path):
     else:
         task_cls = globals()[class_name]
     return task_cls
-
-
-class PSUtilProfiler(psutil.Process):
-    """CPU, I/O and memory profiler using psutil."""
-
-    def __init__(self):
-        self._start_cpu_times = {}
-        self._start_memory = {}
-        self._start_io = {}
-        self._start_disk_io = {}
-        self._start_time = {}
-
-        super().__init__()
-
-        logger.warning(f"Profiling pipeline: {self.cpu_count} cores available.")
-
-    def start(self, name):
-        """
-        Start profiling.
-
-        Results generated when `stop` is called are based on this start time.
-
-        Attributes
-        ----------
-        name : str
-            Description of what is profiled. You have to pass the same str to `stop`.
-        """
-        self._start_time[name] = time.time()
-
-        # Get all stats at the same time
-        with self.oneshot():
-            self._start_cpu_times[name] = self.cpu_times()
-            self.cpu_percent()
-            self._start_memory[name] = self.memory_full_info().uss
-            self._start_disk_io[name] = self.io_counters()
-
-    def stop(self, name):
-        """
-        Stop profiler and return results.
-
-        `start` must be called first.
-
-        Attributes
-        ----------
-        name : str
-            Description of what is profiled. Has to have been passed to `start` before.
-
-        Returns
-        -------
-        cpu_times : `collections.namedtuple`
-            Same as `psutil.cpu_times`. Process CPU times since `start` was called in seconds.
-        cpu_percent :  float
-            Process CPU utilization since `start` was called as percentage. Can be >100 if multiple threads run on
-            different cores. See `PSUtil.cpu_count` for available cores.
-        disk_io : `collections.namedtuple`
-            Same as `psutil.disk_io_counters`. Difference since `start` was called.
-        memory : str
-            Difference of memory in use by this process since `start` was called. If negative,
-            less memory is in use now.
-
-        Raises
-        ------
-        RuntimeError
-            If stop was called before start.
-        """
-        stop_time = time.time()
-
-        # Get all stats at the same time
-        with self.oneshot():
-            cpu_times = self.cpu_times()
-            cpu_percent = self.cpu_percent()
-            disk_io = self.io_counters()
-            memory = self.memory_full_info().uss
-
-        if name not in self._start_cpu_times:
-            raise RuntimeError(
-                f"PSUtilProfiler.stop was called before start for '{name}'."
-            )
-
-        # Construct results
-        CPU_Times = collections.namedtuple("CPUtimes", list(cpu_times._fields))
-        cpu_times = CPU_Times(*np.subtract(cpu_times, self._start_cpu_times.pop(name)))
-        DiskIO = collections.namedtuple("DiskIO", list(disk_io._fields))
-        disk_io = DiskIO(*np.subtract(disk_io, self._start_disk_io.pop(name)))
-        memory = memory - self._start_memory.pop(name)
-        memory = psutil._common.bytes2human(memory)
-        time_s = stop_time - self._start_time.pop(name)
-
-        if time_s < 0.1:
-            logger.warning(
-                f"{name} ran for {time_s:.4f} < 0.1s, results might be inaccurate."
-            )
-
-        logger.info(
-            f"{name} ran for {time_s:.4f}s\n"
-            f"---------------------------------------------------------------------------------------------"
-            f"\n{cpu_times}\n"
-            f"Average CPU load: {cpu_percent}\n"
-            f"{disk_io}\n"
-            f"Change in (uss) memory: {memory}\n"
-            f"============================================================================================="
-        )
-
-    @property
-    def cpu_count(self):
-        return len(self.cpu_affinity())
